+++ server/pkg/key/key.go
@@ -1,54 +1,11 @@
 package key
 
 import (
-	"regexp"
-	"strings"
-
-	"github.com/goombaio/namegenerator"
-	"github.com/reearth/reearthx/util"
-	"github.com/samber/lo"
-	"golang.org/x/exp/slices"
+	"github.com/reearth/reearth-cms/server/pkg/id"
 )
 
-var keyRegexp = regexp.MustCompile("^[a-zA-Z0-9_-]{1,32}$")
+// TODO: completely delete this "key" package
+type Key = id.Key
 
-var ngKeys = []string{"id"}
-
-type Key struct {
-	key string
-}
-
-func New(key string) Key {
-	if !keyRegexp.MatchString(key) {
-		return Key{}
-	}
-	k := Key{key}
-	if !k.IsValid() {
-		return Key{}
-	}
-	return k
-}
-
-func Random() Key {
-	seed := util.Now().UTC().UnixNano()
-	return New(namegenerator.NewNameGenerator(seed).Generate())
-}
-
-func (k Key) IsValid() bool {
-	return k.key != "" && !strings.HasPrefix(k.key, "_") && !strings.HasPrefix(k.key, "-") && !slices.Contains(ngKeys, k.key)
-}
-
-func (k Key) Ref() *Key {
-	return &k
-}
-
-func (k Key) String() string {
-	return k.key
-}
-
-func (k *Key) StringRef() *string {
-	if k == nil {
-		return nil
-	}
-	return lo.ToPtr(k.key)
-}
+var New = id.NewKey
+var Random = id.RandomKey
